+++ JZP_FE/src/pages/JuniorBooking/JuniorBookingHistoryPage.jsx
@@ -1,8 +1,8 @@
 import { useNavigate } from "react-router-dom";
-import Header from "../components/header.jsx";
-import StepBar from "../components/movieStepBar.jsx";
-import HomeIcon from "../assets/icons/homeIcon.svg";
-import BookingCard from "../components/BookingCard";
+import Header from "../../components/header.jsx";
+import StepBar from "../../components/movieStepBar.jsx";
+import HomeIcon from "../../assets/icons/homeIcon.svg";
+import BookingCard from "../../components/BookingCard";
 
 function JuniorBookingHistoryPage() {
   const navigate = useNavigate();
