+++ JZP_FE/src/pages/JuniorBooking/JuniorPayPage.jsx
@@ -1,23 +1,23 @@
 import { useNavigate } from "react-router-dom";
-import Header from "../components/header.jsx";
-import StepBar from "../components/movieStepBar2.jsx";
-import movieImage from "../assets/images/movie2.png";
-import HomeIcon from "../assets/icons/homeIcon.svg";
-import Pay from "../components/Pay2.jsx";
+import Header from "../../components/header.jsx";
+import StepBar from "../../components/movieStepBar.jsx";
+import movieImage from "../../assets/images/movie2.png";
+import HomeIcon from "../../assets/icons/homeIcon.svg";
+import Pay from "../../components/Pay.jsx";
 
-function SeniorPayPage() {
+function JuniorPayPage() {
   const navigate = useNavigate();
 
-  const handleSeniorMain = () => {
-    navigate("/seniorMain");
+  const handleJuniorMain = () => {
+    navigate("/juniorMain");
   };
 
-  const handleSeniorSeat = () => {
-    navigate("/seniorSeat");
+  const handleJuniorSeat = () => {
+    navigate("/juniorSeat");
   };
 
   const handlePayment = () => {
-    navigate("/seniorBooking");
+    navigate("/juniorBooking");
   };
 
   return (
@@ -25,7 +25,7 @@ function SeniorPayPage() {
       <Header />
       <StepBar />
       <div className="flex justify-center py-2">
-        <div className="bg-headerColor text-white p-4 flex flex-col items-center">
+        <div className="bg-buttonGray text-white p-4 flex flex-col items-center">
           {/* 제목 */}
           <h1 className="text-3xl text-sbAggro font-bold text-center mb-4">
             WICKED
@@ -49,34 +49,49 @@ function SeniorPayPage() {
             ></div>
 
             {/* 오른쪽: 영화 정보 */}
-            <div className="flex flex-col justify-start w-2/3 space-y-[28px]">
-              {/* 첫 번째 그룹: 3개 */}
-              <div className="flex flex-col space-y-1">
+            <div className="flex flex-col justify-start w-1/2 space-y-[28px]">
+              {/* 첫 번째 그룹: 2개 */}
+              <div className="flex flex-col space-y-0.5">
+                {[
+                  { label: "영화타입", value: "2D 디지털 더빙" },
+                  { label: "관람등급", value: "15세 이상 이용가" },
+                ].map((item, index) => (
+                  <div key={index} className="flex justify-between text-sm">
+                    <p className="font-semibold text-xs text-textGray mt-0.5">
+                      {item.label}
+                    </p>
+                    <p className="text-[14px] font-medium">{item.value}</p>
+                  </div>
+                ))}
+              </div>
+
+              {/* 두 번째 그룹: 3개 */}
+              <div className="flex flex-col space-y-0.5">
                 {[
                   { label: "상영관", value: "한국공대 2관" },
                   { label: "상영일자", value: "2025.01.06 (월)" },
                   { label: "상영시간", value: "12:30 ~ 14:25" },
                 ].map((item, index) => (
                   <div key={index} className="flex justify-between text-sm">
-                    <p className="font-semibold text-[16px] text-textGray mt-1">
+                    <p className="font-semibold text-xs text-textGray mt-0.5">
                       {item.label}
                     </p>
-                    <p className="text-[18px] font-medium mt-1">{item.value}</p>
+                    <p className="text-[14px] font-medium">{item.value}</p>
                   </div>
                 ))}
               </div>
 
-              {/* 두 번째 그룹: 2개 */}
+              {/* 세 번째 그룹: 2개 */}
               <div className="flex flex-col space-y-0.5">
                 {[
                   { label: "관람인원", value: "성인 1명, 청소년 1명" },
                   { label: "선택좌석", value: "D4, D5" },
                 ].map((item, index) => (
                   <div key={index} className="flex justify-between text-sm">
-                    <p className="font-semibold text-[16px] text-textGray mt-1">
+                    <p className="font-semibold text-xs text-textGray mt-0.5">
                       {item.label}
                     </p>
-                    <p className="text-[18px] font-medium mt-1">{item.value}</p>
+                    <p className="text-[14px] font-medium">{item.value}</p>
                   </div>
                 ))}
               </div>
@@ -92,30 +107,26 @@ function SeniorPayPage() {
         <div className="flex">
           {/* 왼쪽 텍스트 */}
           <div className="w-1/2 pr-4">
-            <p className="text-center text-[22px] font-semibold mt-3 ml-4 whitespace-pre-wrap">
-              결제 금액을
-              <br />
-              확인 해주세요.
+            <p className="text-center text-[14px] font-semibold mt-4 ml-4">
+              결제 금액을 확인하신 후 결제 해주세요.
             </p>
           </div>
 
           {/* 세로 점선 디바이더 */}
-          <div className="border-r border-dashed border-gray-500 mr-3"></div>
+          <div className="border-r border-dashed border-gray-500 mr-4"></div>
 
           {/* 오른쪽 결제 정보 */}
           <div className="w-3/4 mr-8 py-1">
             <div className="flex justify-between mt-2">
-              <span className="text-[16px] text-textGray mt-1 ml-4">성인</span>
+              <span className="text-sm text-textGray mt-1 ml-4">성인</span>
               <span className="text-[18px]">15,000 * 1</span>
             </div>
             <div className="flex justify-between mt-2">
-              <span className="text-[16px] text-textGray mt-1 ml-4">
-                청소년
-              </span>
+              <span className="text-sm text-textGray mt-1 ml-4">청소년</span>
               <span className="text-[18px]">10,000 * 1</span>
             </div>
             <div className="flex justify-between mt-6">
-              <span className="text-[18px] font-bold mt-1 ml-4">
+              <span className="text-[14px] font-bold mt-1 ml-4">
                 총 결제금액
               </span>
               <span className="text-[24px] font-bold text-red-500">
@@ -132,16 +143,16 @@ function SeniorPayPage() {
       {/* Footer 컴포넌트 */}
       <footer className="fixed bottom-0 w-[450px] bg-gray-800 flex mx-auto">
         <button
-          className="flex-1 bg-white text-black text-xl font-bold h-16 flex items-center justify-center leading-none gap-2"
-          onClick={handleSeniorMain}
+          className="flex-1 bg-white text-black text-sm font-bold h-16 flex items-center justify-center leading-none gap-2"
+          onClick={handleJuniorMain}
         >
-          <img src={HomeIcon} alt="홈 아이콘" className="w-8 h-8" />홈
+          <img src={HomeIcon} alt="홈 아이콘" className="w-4 h-4" />홈
         </button>
 
         {/* 결제하기 버튼 */}
         <button
-          className="flex-1 bg-buttonGray text-white text-xl font-bold h-16 flex items-center justify-center leading-none"
-          onClick={handleSeniorSeat}
+          className="flex-1 bg-buttonGray text-white text-sm font-bold h-16 flex items-center justify-center leading-none"
+          onClick={handleJuniorSeat}
         >
           좌석 다시 선택하기
         </button>
@@ -150,4 +161,4 @@ function SeniorPayPage() {
   );
 }
 
-export default SeniorPayPage;
+export default JuniorPayPage;
