+++ JZP_FE/src/pages/JuniorBooking/JuniorSendingPage.jsx
@@ -1,4 +1,4 @@
-import Header from "../components/header.jsx";
+import Header from "../../components/header.jsx";
 
 function JuniorSendingPage() {
   return (
