+++ JZP_FE/src/pages/SeniorBooking/SeniorMovieSelectPage.jsx
@@ -1,52 +1,50 @@
 import { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
-import Header from "../components/header.jsx";
-import StepBar from "../components/movieStepBar.jsx";
-import DateSelectBar from "../components/DateSelectBar.jsx";
-import age12Image from "../assets/images/12.png";
-import age15Image from "../assets/images/15.png";
-import age18Image from "../assets/images/18.png";
-import ageAllImage from "../assets/images/All.png";
-import homeIcon from "../assets/icons/homeIcon.svg";
+import Header from "../../components/header.jsx";
+import StepBar from "../../components/movieStepBar.jsx";
+import DateSelectBar from "../../components/dateSelectBar2.jsx";
+import age12Image from "../../assets/images/12.png";
+import age15Image from "../../assets/images/15.png";
+import age18Image from "../../assets/images/18.png";
+import ageAllImage from "../../assets/images/All.png";
+import homeIcon from "../../assets/icons/homeIcon.svg";
 
 const API_BASE_URL = import.meta.env.VITE_API_BASE_URL;
 
-function JuniorMovieSelectPage() {
+function SeniorMovieSelectPage() {
   const [selectedButton, setSelectedButton] = useState(null);
   const [selectedMovie, setSelectedMovie] = useState(null);
   const [movies, setMovies] = useState([]);
-  const [selectedDate, setSelectedDate] = useState("2025-02-20");
+  const [selectedDate, setSelectedDate] = useState("2025-02-18");
   const navigate = useNavigate();
 
-  const handleJuniorMain = () => navigate("/juniorMain");
+  const handleSeniorMain = () => navigate("/seniorMain");
 
-  const handleJuniorSeatSelect = async () => {
-    if (!selectedMovie) {
-      alert("🎬 영화를 선택해주세요!");
-      return;
-    }
-
-    if (!selectedMovie.movieTime || !selectedMovie.movieTheater) {
-      alert("❗ 영화 시간 또는 상영관 정보가 없습니다. 다시 선택해주세요!");
-      return;
-    }
+  const handleSeniorSeatSelect = async () => {
+    if (!selectedMovie) return;
 
     try {
-      const requestBody = {
-        movieId: selectedMovie.movieId,
-        movieCalendar: selectedDate,
-        movieTime: selectedMovie.movieTime,
-        movieTheater: selectedMovie.movieTheater,
-      };
+      const response = await fetch(`${API_BASE_URL}/api/movie/time`, {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({
+          movieId: selectedMovie.movieId,
+          movieTime: selectedMovie.movieTime,
+          movieTheater: selectedMovie.movieTheater,
+        }),
+      });
 
-      console.log("📤 영화 데이터 저장:", requestBody);
+      const result = await response.json();
+      console.log("🎟️ 좌석 선택 응답:", result);
 
-      // ✅ `localStorage`에 저장 (JuniorSeatSelectPage에서 사용할 데이터)
-      localStorage.setItem("selectedMovie", JSON.stringify(requestBody));
-
-      navigate("/juniorSeat"); // ✅ `state` 없이 이동
+      if (result.status === "success") {
+        navigate("/seniorSeat");
+      } else {
+        alert(`좌석 선택 실패: ${result.message || "알 수 없는 오류"}`);
+      }
     } catch (error) {
-      console.error("🚨 영화 데이터 저장 실패:", error);
+      console.error("🚨 좌석 선택 요청 실패:", error);
+      alert("좌석 선택 요청 중 오류가 발생했습니다.");
     }
   };
 
@@ -87,6 +85,8 @@ function JuniorMovieSelectPage() {
       <StepBar />
       <DateSelectBar onDateChange={setSelectedDate} />
       <div className="h-[1px] bg-gray-700 my-4"></div>
+
+      {/* 영화 리스트 */}
       <div className="flex-1 overflow-y-auto px-4 pb-16">
         {movies.length > 0 ? (
           movies.map((movie, index) => (
@@ -95,7 +95,7 @@ function JuniorMovieSelectPage() {
                 <img
                   src={movie.movieImage}
                   alt={movie.movieName}
-                  className="w-30 h-52 mr-5"
+                  className="w-30 h-52 mr-3"
                 />
                 <div className="flex flex-col justify-between flex-1">
                   <div>
@@ -111,32 +111,32 @@ function JuniorMovieSelectPage() {
                                 : ageAllImage
                         }
                         alt={`${movie.movieRating}세`}
-                        className="w-8 h-8 mr-3"
+                        className="w-10 h-10 mr-3"
                       />
-                      <h3 className="text-xl font-sbAggro font-bold">
+                      <h3 className="text-3xl font-sbAggro font-bold">
                         {movie.movieName}
                       </h3>
                     </div>
-                    <p className="text-sm text-gray-400">
+                    <p className="text-lg text-gray-400">
                       {movie.movieType} | {movie.movieTheater}
                     </p>
                   </div>
-                  <div className="flex flex-wrap gap-2 mt-2">
+                  <div className="grid grid-cols-2 gap-2 mt-2">
                     <button
                       onClick={() => {
                         setSelectedButton(movie.movieId);
                         setSelectedMovie(movie);
                       }}
-                      className={`border w-24 h-12 flex flex-col justify-center items-center ${
+                      className={`border w-36 h-[66px] flex flex-col justify-center items-center ${
                         selectedButton === movie.movieId
                           ? "border-black bg-white text-black"
                           : "border-gray-500 text-white"
                       }`}
                     >
-                      <div className="text-xs font-bold leading-tight">
+                      <div className="text-lg font-bold leading-tight">
                         {movie.movieTime}
                       </div>
-                      <div className="flex justify-between w-full text-[10px] mt-1 px-3">
+                      <div className="flex justify-between w-full text-[14px] mt-1 px-4">
                         <span>{movie.movieSeatRemain}석 남음</span>
                         <span>{movie.movieTheater}</span>
                       </div>
@@ -145,6 +145,7 @@ function JuniorMovieSelectPage() {
                 </div>
               </div>
 
+              {/* 마지막 영화 이후에는 디바이더를 추가하지 않음 */}
               {index !== movies.length - 1 && (
                 <div className="h-[1px] bg-gray-700 my-4"></div>
               )}
@@ -156,16 +157,19 @@ function JuniorMovieSelectPage() {
           </p>
         )}
       </div>
+
+      {/* 하단 네비게이션 */}
       <footer className="fixed bottom-0 w-[450px] bg-gray-800 flex">
         <button
-          className="flex-1 bg-white text-black text-sm font-bold h-16 flex items-center justify-center leading-none gap-2"
-          onClick={handleJuniorMain}
+          className="flex-1 bg-white text-black text-xl font-bold h-20 flex items-center justify-center leading-none gap-2"
+          onClick={handleSeniorMain}
         >
-          <img src={homeIcon} alt="홈 아이콘" className="w-4 h-4" />홈
+          <img src={homeIcon} alt="홈 아이콘" className="w-8 h-8" />홈
         </button>
+
         <button
-          className="flex-1 bg-red-600 text-white text-sm font-bold h-16 flex items-center justify-center leading-none"
-          onClick={handleJuniorSeatSelect}
+          className="flex-1 bg-red-600 text-white text-xl font-bold h-20 flex items-center justify-center leading-none"
+          onClick={handleSeniorSeatSelect}
           disabled={!selectedMovie}
         >
           인원 및 좌석 선택 하기
@@ -175,4 +179,4 @@ function JuniorMovieSelectPage() {
   );
 }
 
-export default JuniorMovieSelectPage;
+export default SeniorMovieSelectPage;
