+++ JZP_FE/src/pages/SeniorBooking/SeniorSendingPage.jsx
@@ -1,5 +1,5 @@
-import Header from "../components/header.jsx";
-import StepBar from "../components/movieStepBar2.jsx";
+import Header from "../../components/header.jsx";
+import StepBar from "../../components/movieStepBar2.jsx";
 
 function SeniorSendingPage() {
   return (
